Move the controller/helper/model overwrite into decorators refer to spree convention
