+++ app/controllers/camaleon_spree/camaleon_cms/frontend_controller_decorator.rb
@@ -1,0 +1,18 @@
+module CamaleonSpree
+  module CamaleonCms
+    module FrontendControllerDecorator
+      def self.prepended(base)
+        base.class_eval do
+          # load basic helpers from spree into camaleon frontend
+          include Spree::Core::ControllerHelpers::Order
+          include Spree::Core::ControllerHelpers::Auth
+          include Spree::Core::ControllerHelpers::Store
+      
+          helper Spree::BaseHelper      
+        end
+      end
+    end
+  end
+end
+
+::CamaleonCms::FrontendController.prepend(CamaleonSpree::CamaleonCms::FrontendControllerDecorator)
